We now always use HTTPS to communicate with Nexmo
