+++ lib/nexmo.js
@@ -10,8 +10,10 @@ The above copyright notice and this permission notice shall be included in all c
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 */
+
+'use strict';
+
 var https = require('https');
-var http = require('http');
 var querystring = require('querystring');
 
 var headers = {
@@ -32,7 +34,6 @@ var controlVerifyEndpoint = {host:'api.nexmo.com',path:'/verify/control/json'};
 var searchVerifyEndpoint = {host:'api.nexmo.com',path:'/verify/search/json'};
 var niEndpoint = {host:'rest.nexmo.com',path:'/ni/json'};
 var up = {};
-var useHttps = false;
 var debugOn = false;
 
 //Error message resources are maintained globally in one place for easy management
@@ -59,8 +60,8 @@ var ERROR_MESSAGES = {
 	numberInsightValidation:'Missing Mandatory fields (number and/or callback url)'
 };
 
-// protocol and debugon are optional
-exports.initialize = function(pkey, psecret, protocol, debugon) {
+// debugon is optional
+exports.initialize = function(pkey, psecret, debugon) {
     if (!pkey || !psecret) {
         throw 'key and secret cannot be empty, set valid values';
     }
@@ -70,7 +71,6 @@ exports.initialize = function(pkey, psecret, protocol, debugon) {
         api_key: pkey,
         api_secret: psecret
     }
-    useHttps = protocol && protocol == 'https'; // default to http
     debugOn = !debugon ? false : (String(debugon) === "false" ? false : true);
     initialized = true;
 }
@@ -200,7 +200,7 @@ function sendRequest(endpoint, method, callback) {
     }
     if (method == 'POST')
         headers['Content-Length'] = 0; // Fix broken due ot 411 Content-Length error now sent by Nexmo API
-    options = {
+    var options = {
         host: endpoint.host?endpoint.host:'rest.nexmo.com',
         port: 80,
         path: '',
@@ -211,12 +211,8 @@ function sendRequest(endpoint, method, callback) {
     log(options);
     var request;
 	if (true) { // set to false to verify the request without sending the actual request
-	    if (useHttps) {
-	        options.port = 443;
-	        request = https.request(options);
-	    } else {
-	        request = http.request(options);
-	    }
+        options.port = 443;
+        request = https.request(options);
 	    request.end();
 	    var responseReturn = '';
 	    request.on('response', function(response) {
